requirements.txt
Twisted >= 16.0.2
PyYAML >= 3.11
MySQL-python >= 1.2.5
